Fixed class naming conventions
