+++ Strings/StringCompression.java
@@ -1,20 +1,22 @@
-public class StringCompression {
-
-	void printCompressedString(String s)
+import java.util.*;
+class StringCompression
+{
+	public static void  main(String args[])
 	{
-		for (int i = 0; i < s.length(); i++) {
-			int count = 1;
-			while (i < s.length() - 1 && s.charAt(i) == s.charAt((i++) + 1))
-				count++;
+		Scanner sc=new Scanner(System.in);
+		String str=sc.nextLine();
+		char ch;int c=0;
+		for(ch='a';ch<='z';ch++)
+		{
+			c=0;
+			for(int i=0;i<str.length();i++)
+			{
+				if(ch==str.charAt(i))
+					c++;
+			}
+			if(c!=0)
+			System.out.print(ch+""+c);
 
-			System.out.print(s.charAt(i));
-			System.out.print(count);
 		}
-	}
-
-	public static void main(String[] args) {
-		StringCompression sc = new StringCompression();
-		String s = "hhhaaaccckkktttooobbbeeerrr";
-		sc.printCompressedString(s);
 	}
 }
